requirements.txt
-e git+https://github.com/django/django.git#egg=django1.5-dev
django-tagging==0.3.1
sorl-thumbnail==11.12
south==0.7.4